Removed Consolelogs Removed apiservice class Added Error Handler Service
