+++ src/app/classes/apiresponse.ts
@@ -1,0 +1,5 @@
+export interface APIResponse<D> {
+    code: number,
+    msg: string,
+    data: D
+}
