+++ src/app/services/auth.service.ts
@@ -1,22 +1,25 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { catchError, Observable, share, take, tap } from 'rxjs';
-import { APIResponse, ApiService } from './apiservice';
-import { NotificationService } from './notification.service';
+import { catchError, take, tap } from 'rxjs';
+import { environment } from 'src/environments/environment';
+import { APIResponse } from '../classes/apiresponse';
+import { ErrorHandlerService } from './error-handler.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class AuthService extends ApiService{
-  authUrl = this.apiUrl + '/auth'
+export class AuthService {
+  authUrl = environment.baseUrl + '/auth'
 
-  constructor(private http: HttpClient, notificationService: NotificationService) {
-    super(notificationService);
-  }
+  constructor(
+    private http: HttpClient,
+    private errorHandler: ErrorHandlerService
+  ) {}
 
-  auth(username: string, password: string) :Observable<APIResponse<string>> {
-    console.log("AUTH REQUEST")
-    const options = {headers: this.headers}
-    return this.http.post<APIResponse<string>>(this.authUrl, {"username":username,"password":password}, options)
+  auth(username: string, password: string) {
+    const options = {headers: environment.requestHeaders}
+    return this.http.post<APIResponse<string>>(this.authUrl, {"username":username,"password":password}, options).pipe(
+      catchError(this.errorHandler.handleError)
+    )
   }
 }
