+++ src/app/state/auth/auth.effects.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core"
 import { Router } from "@angular/router"
 import { Actions, createEffect, ofType } from "@ngrx/effects"
-import { catchError, concatMap, exhaustMap, map, of, switchMap, tap } from "rxjs"
+import { catchError, map, of, switchMap, tap } from "rxjs"
 import { AuthService } from "src/app/services/auth.service"
 import { CookieService } from "src/app/services/cookie.service"
 import { AuthActions } from "./auth.actions"
@@ -15,22 +15,18 @@ export class AuthEffects {
         private authService: AuthService,
         private cookieService: CookieService,
         private router: Router
-    ) {
-        console.log("AUTH EFFECTS")
-    }
+    ) {}
 
 
     auth$ = createEffect(() => 
         this.actions$.pipe(
             ofType(AuthActions.authLogin),
-            tap(() => console.log("AUTH ACTION")),
             switchMap(authAction => 
                 this.authService.auth(authAction.username, authAction.password).pipe(
-                    tap(response => {console.log(response.data)}),
                     map(response => AuthActions.authSuccess({token: response.data})),      
+                    catchError(error => of(AuthActions.authFailure({ message: error }))),
                 )
-            ),
-            tap(console.log)
+            )
         )
     )
