+++ src/environments/environment.prod.ts
@@ -1,5 +1,7 @@
+import { HttpHeaders } from "@angular/common/http";
+
 export const environment = {
   production: true,
-  baseUrl: 'https://stratosphaere.codelix.de/api/',
-  apiVersion: 'v1',
+  requestHeaders: new HttpHeaders({'Content-Type' : 'application/json'}),
+  baseUrl: 'https://stratosphaere.codelix.de/api/v1',
 };
